Telomare

Documentation

class BasicBase g where

Methods

embedB :: BasicExprF x -> g x

extractB :: g x -> Maybe (BasicExprF x)

class StuckBase g where

Methods

embedS :: StuckF x -> g x

extractS :: g x -> Maybe (StuckF x)

class AbortBase g where

Methods

embedA :: AbortableF x -> g x

extractA :: g x -> Maybe (AbortableF x)

pattern BasicFW :: BasicBase g => BasicExprF x -> g x

pattern BasicEE :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => BasicExprF g -> g

pattern StuckFW :: StuckBase g => StuckF x -> g x

pattern StuckEE :: (Base g ~ f, StuckBase f, Recursive g, Corecursive g) => StuckF g -> g

pattern AbortFW :: AbortBase g => AbortableF x -> g x

pattern AbortEE :: (Base g ~ f, AbortBase f, Recursive g, Corecursive g) => AbortableF g -> g

pattern FillFunction :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> f g

pattern FillFunctionEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g

pattern GateSwitch :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g -> f g

pattern GateSwitchEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g -> g

pattern AppEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g

pattern EnvB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g

pattern SetEnvB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g

pattern GateB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g -> g

pattern LeftB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g

pattern RightB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g

pattern ZeroB :: (Recursive g, Corecursive g, Base g ~ f, BasicBase f) => g

pattern AbortB :: (Recursive g, Corecursive g, Base g ~ f, AbortBase f) => g

pattern PairB :: (Recursive g, Corecursive g, Base g ~ f, BasicBase f) => g -> g -> g

varB :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => Int -> g

i2B :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => Int -> g

b2i :: (Base g ~ f, BasicBase f, Recursive g) => g -> Maybe Int

b2s :: forall g f. (Base g ~ f, CarryWrap g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g) => g -> Maybe String

s2b :: forall g f w. (Base g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => String -> g

iteB_ :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => g -> g -> g -> g

data BasicExprF f

Constructors

ZeroSF

PairSF f f

type BasicExpr = Fix BasicExprF

data StuckF f

Constructors

DeferSF FunctionIndex f

EnvSF

SetEnvSF f

GateSF f f

LeftSF f

RightSF f

newtype FunctionIndex

Constructors

FunctionIndex

Fields

unFunctionIndex :: Int

data AbortableF f

Constructors

AbortF

AbortedF BasicExpr

data StuckExprF f

Constructors

StuckExprB (BasicExprF f)

StuckExprS (StuckF f)

type StuckExpr = Fix StuckExprF

data CompiledExprF f

Constructors

CompiledExprB (BasicExprF f)

CompiledExprS (StuckF f)

CompiledExprA (AbortableF f)

type CompiledExpr = Fix CompiledExprF

newtype UnsizedRecursionToken

Constructors

UnsizedRecursionToken

Fields

unUnsizedRecursionToken :: Int

data SourcePosition

Constructors

SourcePosition

Fields

sourcePositionLine :: Int

sourcePositionColumn :: Int

sourcePositionOffset :: Int

data SourceSpan

Constructors

SourceSpan

Fields

sourceSpanFile :: Maybe FilePath

sourceSpanStart :: SourcePosition

sourceSpanEnd :: SourcePosition

data LocTag

Constructors

SourceLoc SourceSpan

GeneratedLoc String (Maybe LocTag)

BuiltinLoc String

RuntimeLoc

DecompiledLoc

UnknownLoc

data Term3F f

Constructors

Term3B (BasicExprF f)

Term3S (StuckF f)

Term3A (AbortableF f)

Term3Unsized UnsizedRecursionToken

Term3CheckingWrapper LocTag f f

data PatternF t f

AST for patterns in `case` expressions

Constructors

PatternVarF String

PatternAnnotatedF f t

PatternIntF Int

PatternStringF String

PatternIgnoreF

PatternPairF f f

data UnprocessedParsedTermF p f

Firstly parsed AST

Constructors

UnprocessedParsedTermH (HighTermF f)

UnprocessedParsedTermL (LamTermF LocatedName String f)

UnprocessedParsedTermB (BasicExprF f)

LetUPF [(LocatedName, f)] f

ListUPF [f]

IntUPF Int

StringUPF String

UDTUPF [String] f

CaseUPF f [(p, f)]

ImportQualifiedUPF String String

ImportUPF String

type Pattern = Fix (PatternF UnprocessedParsedTerm)

newtype UnprocessedParsedTerm

Constructors

UnprocessedParsedTerm

Fields

unUnprocessedParsedTerm :: UPT

type UPT = Fix (UnprocessedParsedTermF Pattern)

data LamType l

Lambdas can be closed if it's expresion does not depend on any outer binding.

Constructors

Open l

Closed l

LetBinding Int l

class LamBase g where

Associated Types

type LamVar g

type LamT g

Methods

embedL :: LamTermF (LamT g) (LamVar g) x -> g x

extractL :: g x -> Maybe (LamTermF (LamT g) (LamVar g) x)

class HighBase g where

Methods

embedH :: HighTermF x -> g x

extractH :: g x -> Maybe (HighTermF x)

data LamTermF l v f

Constructors

VarF v

AppF f f

LamF l f

data HighTermF f

High level grammar elements

Constructors

CheckF f f

ITEF f f f

HLeftF f

HRightF f

HTraceF f

HashF f
On ad hoc user defined types, this term will be substitued to a unique Int.

ChurchF Int

RecursionF f f f

newtype AnnotatedUPT

Constructors

AnnotatedUPT

Fields

unAnnotatedUPT :: AUPT

type AUPT = Cofree (UnprocessedParsedTermF PatternA) LocTag

type PatternA = Fix (PatternF AnnotatedUPT)

data ParserTermF l v f

Parser AST

Constructors

ParserTermB (BasicExprF f)

ParserTermH (HighTermF f)

ParserTermL (LamTermF l v f)

TUnsizedRepeaterF

indent :: Int -> String -> String

Helper function to indent. Usefull for indented Show instances.

sindent :: String -> State Int String

Indentation with the State Monad.

indentWithOneChild :: String -> State Int String -> State Int String

One child indentation.

indentWithOneChild' :: String -> State Int String -> State Int String

indentWithTwoChildren' :: String -> State Int String -> State Int String -> State Int String

indentWithChildren' :: String -> [State Int String] -> State Int String

indentWithTwoChildren :: String -> State Int String -> State Int String -> State Int String

Two children indentation.

indentWithThreeChildren :: String -> State Int String -> State Int String -> State Int String -> State Int String

dropUntil :: (a -> Bool) -> [a] -> [a]

`dropUntil p xs` drops leading elements until `p $ head xs` is satisfied.

locStartLineColumn :: LocTag -> Maybe (Int, Int)

renderLocTag :: LocTag -> Maybe String

renderLocTagCompact :: LocTag -> Maybe String

`file:line:column` when the location knows which file it came from, `line:column` when it doesn't. Compact enough to list many of them.

renderLocTagVerbose :: LocTag -> String

type Term1 = Cofree (ParserTermF (LamType String) String) LocTag

type Term2 = Cofree (ParserTermF (LamType ()) Int) LocTag

type Term3 = Cofree Term3F LocTag

data TypeCheckError

Constructors

UnboundType Int

InconsistentTypes PartialType PartialType

RecursiveType Int

data ResolverError

Constructors

NoMainFunction String

ModuleNotFound String

DefinitionCycle [String]

MissingDefinitions [String]

MissingDefinitionAt LocTag String

ParseError String

renderResolverError :: ResolverError -> String

data SizingFailureKind

Why the sizing pass could not pin a finite iteration count on a recursion site. The two cases need different advice, so they are kept apart.

Constructors

FuelExhausted Int
The abstract unrolling reached this depth and the recursion's test still had not said "stop".

UnboundedInput
The recursion's test reached input that no static refinement bounds, so no unrolling depth at all can be proven sufficient.

data SizingFailure

A sizing failure, named well enough to act on: which recursion, why, and where it is in the source.

Constructors

SizingFailure

Fields

sizingFailureToken :: UnsizedRecursionToken

sizingFailureKind :: SizingFailureKind

sizingFailureLoc :: Maybe LocTag

renderSizingSite :: SizingFailure -> String

Names the recursion site as precisely as the available information allows.

renderSizingFailure :: SizingFailure -> String

data EvalError

Constructors

RTE RunTimeError

TCE TypeCheckError

RE ResolverError

StaticCheckError String

CompileConversionError

RecursionLimitError SizingFailure

renderEvalError :: EvalError -> String

What to show a user. Only a sizing failure reads differently from show: it has advice worth giving unadorned.

data RunTimeError

Constructors

AbortRunTime BasicExpr

GenericRunTimeError String CompiledExpr

ResultConversionError String

class TelomareLike a where

Methods

fromTelomare :: StuckExpr -> a

toTelomare :: a -> Maybe StuckExpr

class TelomareLike a => AbstractRunTime a where

Methods

eval :: a -> Either RunTimeError a

type Term3Builder g = State (FunctionIndex, UnsizedRecursionToken) g

buildTerm :: Corecursive g => Term3Builder g -> g

deferS :: (Base g ~ f, StuckBase f, Recursive g, Corecursive g) => g -> Term3Builder g

pairS :: (Base g ~ CofreeF f a, BasicBase f, Recursive g, Corecursive g, Monad m) => m g -> m g -> m g

clamS :: forall g f. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g) => Term3Builder g -> Term3Builder g

lamS :: forall g f. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g) => Term3Builder g -> Term3Builder g

twiddleS :: forall g f w. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => Term3Builder g

appS :: forall g f w. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => Term3Builder g -> Term3Builder g -> Term3Builder g

repeatFunctionS :: LocTag -> Term3Builder Term3

unsizedRepeater :: LocTag -> UnsizedRecursionToken -> Term3Builder Term3

repeaterAndAbort :: LocTag -> UnsizedRecursionToken -> Term3Builder Term3

i2CB :: LocTag -> Int -> Term3Builder Term3

unsizedRecursionWrapper :: LocTag -> Term3Builder Term3 -> Term3Builder Term3 -> Term3Builder Term3 -> Term3Builder Term3

data DataType

Constructors

ZeroType

ArrType DataType DataType

PairType DataType DataType

data PartialTypeF f

Constructors

ZeroTypeP

AnyType

TypeVariable LocTag Int

ArrTypeP f f

PairTypeP f f

type PartialType = Fix PartialTypeF

toPartialType :: DataType -> PartialType

mergePairType :: DataType -> DataType

mergePairTypeP :: PartialType -> PartialType

containsFunction :: PartialType -> Bool

cleanType :: PartialType -> Bool

forget :: Corecursive a => Cofree (Base a) anno -> a

tag :: Recursive a => anno -> a -> Cofree (Base a) anno

convertBasic :: (BasicBase g, BasicBase h, Base x ~ h, Recursive x, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

convertStuck :: (StuckBase g, StuckBase h, Base x ~ h, Recursive x, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

convertAbort :: (AbortBase g, AbortBase h, Base x ~ h, Recursive x, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

insertAndGetKey :: (Ord e, Enum e) => a -> State (Map e a) e

pattern AbortRecursion :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g -> g

pattern AbortUser :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g -> g

pattern AbortAny :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g

pattern AbortUnsizeable :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g -> g

convertAbortMessage :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, Show g) => g -> String

newtype LocatedName

Constructors

LocatedName (LocTag, String)

locatedNameLoc :: LocatedName -> LocTag

locatedNameText :: LocatedName -> String

locatedName :: LocTag -> String -> LocatedName

letBindingName :: (LocatedName, a) -> String

letBindingValue :: (LocatedName, a) -> a

letBindingLoc :: (LocatedName, a) -> LocTag

class CarryAnno g where

Associated Types

type CarryWrap g :: * -> *

Methods

getEmbed :: g -> CarryWrap g g -> g

pattern GFix :: (Recursive g, Corecursive g, Base g ~ f) => f g -> g

pattern VarFP :: (LamBase f, LamVar f ~ n) => n -> f x

pattern VarAFP :: (LamBase f, LamVar f ~ n) => a -> n -> CofreeF f a x

pattern VarP :: (Recursive g, Corecursive g, Base g ~ f, LamBase f, LamVar f ~ n) => n -> g

pattern AppAFP :: LamBase f => a -> b -> b -> CofreeF f a b

pattern AppP :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, LamBase w, LamBase f) => g -> g -> g

pattern LamAFP :: (LamBase f, LamT f ~ n) => a -> n -> b -> CofreeF f a b

pattern LamP :: (Recursive g, CarryAnno g, Base g ~ f, LamBase f, CarryWrap g ~ w, LamBase w, LamT w ~ n, LamT f ~ n) => n -> g -> g

pattern ITEAFP :: HighBase f => LocTag -> b -> b -> b -> CofreeF f LocTag b

pattern ITEP :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g -> g -> g

pattern HLeft :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g

pattern HRight :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g

pattern HTrace :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g

pattern PairAFP :: BasicBase f => LocTag -> x -> x -> CofreeF f LocTag x

pattern PairP :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, BasicBase f, BasicBase w) => g -> g -> g