Telomare.IR.Base

Description

The shared functor vocabulary of the compiler. Telomare's IRs are not independent grammars: each one is a composition of the base functors defined here (BasicExprF, StuckF, AbortableF, LamTermF, HighTermF), glued together by the *Base embed/extract classes. The pattern synonyms are polymorphic over those classes, so one synonym (e.g. PairB) works at every stage of the pipeline.

Documentation

class BasicBase g where

Methods

embedB :: BasicExprF x -> g x

extractB :: g x -> Maybe (BasicExprF x)

class StuckBase g where

Methods

embedS :: StuckF x -> g x

extractS :: g x -> Maybe (StuckF x)

class AbortBase g where

Methods

embedA :: AbortableF x -> g x

extractA :: g x -> Maybe (AbortableF x)

pattern BasicFW :: BasicBase g => BasicExprF x -> g x

pattern BasicEE :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => BasicExprF g -> g

pattern StuckFW :: StuckBase g => StuckF x -> g x

pattern StuckEE :: (Base g ~ f, StuckBase f, Recursive g, Corecursive g) => StuckF g -> g

pattern AbortFW :: AbortBase g => AbortableF x -> g x

pattern AbortEE :: (Base g ~ f, AbortBase f, Recursive g, Corecursive g) => AbortableF g -> g

pattern FillFunction :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> f g

pattern FillFunctionEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g

pattern GateSwitch :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g -> f g

pattern GateSwitchEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g -> g

pattern AppEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g

pattern EnvB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g

pattern SetEnvB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g

pattern GateB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g

pattern LeftB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g

pattern RightB :: (Recursive g, Corecursive g, Base g ~ f, StuckBase f) => g -> g

pattern ZeroB :: (Recursive g, Corecursive g, Base g ~ f, BasicBase f) => g

pattern AbortB :: (Recursive g, Corecursive g, Base g ~ f, AbortBase f) => g

pattern PairB :: (Recursive g, Corecursive g, Base g ~ f, BasicBase f) => g -> g -> g

varB :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => Int -> g

i2B :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => Int -> g

b2i :: (Base g ~ f, BasicBase f, Recursive g) => g -> Maybe Int

b2s :: forall g f. (Base g ~ f, CarryWrap g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g) => g -> Maybe String

s2b :: forall g f w. (Base g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => String -> g

iteB_ :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g, CarryAnno g, CarryWrap g ~ w, BasicBase w) => g -> g -> g -> g

data BasicExprF f

Constructors

ZeroSF

PairSF f f

type BasicExpr = Fix BasicExprF

data StuckF f

Constructors

DeferSF FunctionIndex f

EnvSF

SetEnvSF f

GateSF

LeftSF f

RightSF f

newtype FunctionIndex

Constructors

FunctionIndex

Fields

unFunctionIndex :: Int

data AbortableF f

Constructors

AbortF

AbortedF BasicExpr

newtype UnsizedRecursionToken

Constructors

UnsizedRecursionToken

Fields

unUnsizedRecursionToken :: Int

data LamType l

Lambdas can be closed if it's expresion does not depend on any outer binding.

Constructors

Open l

Closed l

LetBinding Int l

class LamBase g where

Associated Types

type LamVar g

type LamT g

Methods

embedL :: LamTermF (LamT g) (LamVar g) x -> g x

extractL :: g x -> Maybe (LamTermF (LamT g) (LamVar g) x)

class HighBase g where

Methods

embedH :: HighTermF x -> g x

extractH :: g x -> Maybe (HighTermF x)

data LamTermF l v f

Constructors

VarF v

AppF f f

LamF l f

data HighTermF f

High level grammar elements

Constructors

CheckF f f

ITEF f f f

HLeftF f

HRightF f

HTraceF f

HashF f
On ad hoc user defined types, this term will be substitued to a unique Int.

ChurchF Int

RecursionF f f f

forget :: Corecursive a => Cofree (Base a) anno -> a

tag :: Recursive a => anno -> a -> Cofree (Base a) anno

convertBasic :: (BasicBase g, BasicBase h, Base x ~ h, Recursive x, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

convertStuck :: (StuckBase g, StuckBase h, Base x ~ h, Recursive x, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

convertAbort :: (AbortBase g, AbortBase h, Base x ~ h, Recursive x, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

insertAndGetKey :: (Ord e, Enum e) => a -> State (Map e a) e

pattern AbortRecursion :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g -> g

pattern AbortUser :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g -> g

pattern AbortAny :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g

pattern AbortUnsizeable :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, CarryAnno g, Recursive g, Corecursive g) => g -> g

convertAbortMessage :: (Base g ~ f, CarryWrap g ~ f, BasicBase f, Recursive g, Corecursive g, CarryAnno g, Show g) => g -> String

class CarryAnno g where

Associated Types

type CarryWrap g :: Type -> Type

Methods

getEmbed :: g -> CarryWrap g g -> g

pattern GFix :: (Recursive g, Corecursive g, Base g ~ f) => f g -> g

pattern VarFP :: (LamBase f, LamVar f ~ n) => n -> f x

pattern VarAFP :: (LamBase f, LamVar f ~ n) => a -> n -> CofreeF f a x

pattern VarP :: (Recursive g, Corecursive g, Base g ~ f, LamBase f, LamVar f ~ n) => n -> g

pattern AppAFP :: LamBase f => a -> b -> b -> CofreeF f a b

pattern AppP :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, LamBase w, LamBase f) => g -> g -> g

pattern LamAFP :: (LamBase f, LamT f ~ n) => a -> n -> b -> CofreeF f a b

pattern LamP :: (Recursive g, CarryAnno g, Base g ~ f, LamBase f, CarryWrap g ~ w, LamBase w, LamT w ~ n, LamT f ~ n) => n -> g -> g

pattern ITEAFP :: HighBase f => LocTag -> b -> b -> b -> CofreeF f LocTag b

pattern ITEP :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g -> g -> g

pattern HLeft :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g

pattern HRight :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g

pattern HTrace :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, HighBase f, HighBase w) => g -> g

pattern PairAFP :: BasicBase f => LocTag -> x -> x -> CofreeF f LocTag x

pattern PairP :: (Recursive g, CarryAnno g, Base g ~ f, CarryWrap g ~ w, BasicBase f, BasicBase w) => g -> g -> g