Telomare.IR.Loc

Description

Source locations. The lowest layer of the IR vocabulary: every stage annotates its terms with a LocTag so diagnostics can say where a term came from.

Documentation

data SourcePosition

Constructors

SourcePosition

Fields

sourcePositionLine :: Int

sourcePositionColumn :: Int

sourcePositionOffset :: Int

data SourceSpan

Constructors

SourceSpan

Fields

sourceSpanFile :: Maybe FilePath

sourceSpanStart :: SourcePosition

sourceSpanEnd :: SourcePosition

data LocTag

Constructors

SourceLoc SourceSpan

GeneratedLoc String (Maybe LocTag)

BuiltinLoc String

RuntimeLoc

DecompiledLoc

UnknownLoc

locStartLineColumn :: LocTag -> Maybe (Int, Int)

renderLocTag :: LocTag -> Maybe String

renderLocTagCompact :: LocTag -> Maybe String

`file:line:column` when the location knows which file it came from, `line:column` when it doesn't. Compact enough to list many of them.

renderLocTagVerbose :: LocTag -> String

newtype LocatedName

Constructors

LocatedName (LocTag, String)

locatedNameLoc :: LocatedName -> LocTag

locatedNameText :: LocatedName -> String

locatedName :: LocTag -> String -> LocatedName

letBindingName :: (LocatedName, a) -> String

letBindingValue :: (LocatedName, a) -> a

letBindingLoc :: (LocatedName, a) -> LocTag