Telomare.Possible

Documentation

debug :: Bool

debugTrace :: String -> a -> a

anaM' :: (Monad m, Corecursive t, x ~ Base t, Traversable x) => (a -> m (Base t a)) -> a -> m t

basicStep :: (Base g ~ f, BasicBase f, Corecursive g, Recursive g) => (f g -> g) -> f g -> g

basicStepM :: (Base g ~ f, BasicBase f, Traversable f, Corecursive g, Recursive g, PrettyPrintable g, Monad m) => (f g -> m g) -> f g -> m g

transformNoDefer :: (Base g ~ f, StuckBase f, Recursive g) => (f g -> g) -> g -> g

transformNoDeferM :: (Base g ~ f, StuckBase f, Traversable f, Monad m, Recursive g) => (f g -> m g) -> g -> m g

stuckStep :: (Base a ~ f, StuckBase f, BasicBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

stuckStepDebug :: (Base a ~ f, StuckBase f, BasicBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

stuckStepM :: (Base a ~ f, Traversable f, StuckBase f, BasicBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => (f a -> m a) -> f a -> m a

stuckStepDebugM :: forall a f m j. (Base a ~ f, Traversable f, StuckBase f, BasicBase f, AbortBase f, UnsizedBase f, IndexedInputBase f, SuperBase f, ShallowEq1 f, Show a, Eq a, Recursive a, Corecursive a, PrettyPrintable a, GenValidAdj a, Monad m, Gennable a ~ j, GenValid j, Eq j) => (f a -> m a) -> f a -> m a

stuckStepWithTrace :: (Base a ~ f, MonadReader s m, s ~ TCallStack a, Traversable f, StuckBase f, BasicBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> m a) -> f a -> m a

failAndPrintStack :: (Base a ~ f, MonadReader s m, s ~ TCallStack a, Corecursive a, PrettyPrintable a) => f a -> m b

gateBasicResult :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => (g -> GateResult g) -> g -> GateResult g

gateSuperResult :: (Base g ~ f, SuperBase f, Recursive g, Corecursive g) => (g -> GateResult g) -> (g -> GateResult g) -> g -> GateResult g

gateAbortResult :: (Base g ~ f, AbortBase f, Recursive g, Corecursive g) => (g -> GateResult g) -> g -> GateResult g

gateIndexedResult :: (Base g ~ f, IndexedInputBase f, Recursive g, Corecursive g) => (g -> GateResult g) -> g -> GateResult g

mergeShallow :: (Base g ~ f, SuperBase f, ShallowEq1 f, Recursive g, Corecursive g) => Maybe Integer -> g -> g -> g

foldGateResult :: forall g f. (Base g ~ f, SuperBase f, Recursive g, Corecursive g) => Maybe Integer -> g -> g -> GateResult g -> g

superStep :: forall a f. (Base a ~ f, BasicBase f, SuperBase f, ShallowEq1 f, Recursive a, Corecursive a, PrettyPrintable a) => (a -> GateResult a) -> (f a -> a) -> (f a -> a) -> f a -> a

superUnsizedStep :: forall a f. (Base a ~ f, Traversable f, BasicBase f, SuperBase f, UnsizedBase f, ShallowEq1 f, Recursive a, Corecursive a, PrettyPrintable a) => (a -> GateResult a) -> (f a -> a) -> (f a -> a) -> f a -> a

superStepM :: forall a f m. (Base a ~ f, Traversable f, BasicBase f, SuperBase f, ShallowEq1 f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => (a -> GateResult a) -> (f a -> m a) -> (f a -> m a) -> f a -> m a

superStepM' :: forall a f m. (Base a ~ f, Traversable f, BasicBase f, SuperBase f, ShallowEq1 f, Recursive a, Corecursive a, PrettyPrintable a, Show a, Monad m) => (a -> GateResult a) -> (f a -> m a) -> (f a -> m a) -> f a -> m a

superAbortStep :: (Base g ~ f, Traversable f, BasicBase f, SuperBase f, AbortBase f, ShallowEq1 f, Recursive g, Corecursive g) => (f g -> g) -> (f g -> g) -> f g -> g

superAbortStepM :: (Base g ~ f, Traversable f, BasicBase f, SuperBase f, AbortBase f, ShallowEq1 f, Recursive g, Corecursive g, Monad m) => (f g -> m g) -> (f g -> m g) -> f g -> m g

indexedAbortStep :: (Base a ~ f, Traversable f, BasicBase f, AbortBase f, IndexedInputBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

indexedAbortStepM :: (Base a ~ f, Traversable f, BasicBase f, AbortBase f, IndexedInputBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => (f a -> m a) -> f a -> m a

indexedSuperStep :: (Base a ~ f, Traversable f, BasicBase f, SuperBase f, IndexedInputBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

indexedSuperStepM :: (Base a ~ f, Traversable f, BasicBase f, SuperBase f, IndexedInputBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => (f a -> m a) -> f a -> m a

fuzzyStepM :: (Base a ~ f, Traversable f, BasicBase f, FuzzyBase f, Recursive a, Corecursive a, Show a, PrettyPrintable a, Monad m) => (a -> a -> a) -> (f (m a) -> m a) -> (f (m a) -> m a) -> f (m a) -> m a

abortStep :: (Base a ~ f, BasicBase f, StuckBase f, AbortBase f, Recursive a, Corecursive a) => (f a -> a) -> f a -> a

abortStepM :: (Base a ~ f, Traversable f, BasicBase f, StuckBase f, AbortBase f, Recursive a, Corecursive a, Monad m) => (f a -> m a) -> f a -> m a

twiddleInd :: Int

deferB :: (Base g ~ f, StuckBase f, Recursive g, Corecursive g) => Int -> g -> g

lamB :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => Int -> g -> g

twiddleB :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g

appB :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g

iteB :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g -> g

argOneB :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => g

argTwoB :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => g

argThreeB :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => g

argFourB :: (Base g ~ f, BasicBase f, Recursive g, Corecursive g) => g

unsizedTestIndexed :: (Base g ~ f, BasicBase f, AbortBase f, IndexedInputBase f, Recursive g, Corecursive g) => Set Integer -> (UnsizedRecursionToken -> g -> g) -> UnsizedRecursionToken -> g -> g

unsizedTestSuper :: (Base g ~ f, SuperBase f, AbortBase f, Recursive g, Corecursive g) => (g -> g) -> (UnsizedRecursionToken -> g -> g) -> UnsizedRecursionToken -> g -> g

unsizedTestDeferred :: (Base g ~ f, DeferredEvalBase f, Recursive g, Corecursive g) => (UnsizedRecursionToken -> g -> g) -> UnsizedRecursionToken -> g -> g

unsizedTestUnsized :: (Base g ~ f, UnsizedBase f, Recursive g, Corecursive g) => (g -> g) -> (UnsizedRecursionToken -> g -> g) -> UnsizedRecursionToken -> g -> g

unsizedStep :: forall a f. (Base a ~ f, Traversable f, BasicBase f, StuckBase f, AbortBase f, UnsizedBase f, Recursive a, Corecursive a, Eq a, PrettyPrintable a) => Int -> (UnsizedRecursionToken -> a -> a) -> (f a -> a) -> (f a -> a) -> f a -> a

unsizedStepM :: forall a f t m. (Base a ~ f, Traversable f, BasicBase f, StuckBase f, AbortBase f, UnsizedBase f, Recursive a, Corecursive a, Eq a, PrettyPrintable a, m ~ StrictAccum SizedRecursion, MonadTrans t, Applicative (t m)) => Int -> (UnsizedRecursionToken -> a -> a) -> (f a -> t m a) -> f a -> t m a

forceSizes :: Int -> UnsizedExpr -> UnsizedExpr

unsizedStepM' :: forall a f t m. (Base a ~ f, Traversable f, BasicBase f, StuckBase f, AbortBase f, UnsizedBase f, Recursive a, Corecursive a, Eq a, PrettyPrintable a, m ~ StrictAccum SizedRecursion, MonadTrans t, Applicative (t m)) => Int -> Set Integer -> (UnsizedRecursionToken -> a -> a) -> (f a -> t m a) -> f a -> t m a

unsizedStepM''' :: forall a f t m. (Base a ~ f, Traversable f, BasicBase f, StuckBase f, AbortBase f, UnsizedBase f, Recursive a, Corecursive a, Eq a, PrettyPrintable a, m ~ StrictAccum SizedRecursion, MonadTrans t, Applicative (t m)) => Int -> Set Integer -> (UnsizedRecursionToken -> a -> a) -> (f a -> t m a) -> f a -> t m a

zeroedInputStepM :: (Base a ~ g, Traversable f, IndexedInputBase f, BasicBase g, Corecursive a, Monad m) => Set Integer -> (f a -> m a) -> f a -> m a

indexedInputStep :: (Base a ~ f, BasicBase f, IndexedInputBase f, Recursive a, Corecursive a) => Set Integer -> (f a -> a) -> f a -> a

indexedInputStep' :: (Base a ~ f, BasicBase f, IndexedInputBase f, Recursive a, Corecursive a) => Set Integer -> (f a -> a) -> f a -> a

indexAbortIfUnboundStep :: (Base a ~ f, BasicBase f, AbortBase f, IndexedInputBase f, Recursive a, Corecursive a, Show a) => Set Integer -> (f a -> a) -> f a -> a

indexedInputStepM :: (Base a ~ f, Traversable f, BasicBase f, StuckBase f, IndexedInputBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => Set Integer -> (f a -> m a) -> f a -> m a

indexedInputStepM' :: (Base a ~ f, Traversable f, BasicBase f, StuckBase f, IndexedInputBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => Set Integer -> (f a -> m a) -> f a -> m a

indexedInputIgnoreSwitchStepM :: (Base a ~ f, Traversable f, BasicBase f, IndexedInputBase f, Recursive a, Corecursive a, Monad m) => (f a -> m a) -> f a -> m a

indexSwitchSuperSplitStep :: (Base a ~ f, BasicBase f, IndexedInputBase f, SuperBase f, Recursive a, Corecursive a) => (f a -> a) -> f a -> a

deferredEvalStep :: (Base a ~ f, Traversable f, BasicBase f, DeferredEvalBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

deferredEvalStep' :: (Base a ~ f, Traversable f, BasicBase f, DeferredEvalBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

abortDeferredStep :: (Base a ~ f, BasicBase f, AbortBase f, DeferredEvalBase f, Recursive a, Corecursive a) => (f a -> a) -> f a -> a

isUnbounded :: Set Integer -> Integer -> Bool

extractZeroes :: InputSizingExpr -> Set Integer

zeroToBranch :: (Base a ~ f, BasicBase f, Corecursive a) => Integer -> a

testB :: UnsizedExpr

pathToBranch :: Integer -> String

findInputLimitStepM :: (InputSizingExprF InputSizingExpr -> StrictAccum (Set Integer) InputSizingExpr) -> InputSizingExprF InputSizingExpr -> StrictAccum (Set Integer) InputSizingExpr

unsized2abortExpr :: UnsizedExpr -> AbortExpr

term3ToUnsizedExpr :: Int -> Term3 -> UnsizedExpr

getInputLimits :: UnsizedExpr -> Set Integer

capMain :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> g -> g

isClosure :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g, Corecursive g) => g -> Bool

sizeTerm :: Int -> UnsizedExpr -> Either UnsizedRecursionToken AbortExpr

sizeTermM :: Int -> UnsizedExpr -> Either UnsizedRecursionToken AbortExpr

abortPossibilities :: Int -> UnsizedExpr -> Set IExpr

getSizesM :: Int -> UnsizedExpr -> Either UnsizedRecursionToken SizedRecursion

buildUnsizedLocMap :: UnsizedExpr -> Map UnsizedRecursionToken LocTag

removeRefinementWrappers :: (Base g ~ f, BasicBase f, StuckBase f, AbortBase f, UnsizedBase f, Recursive g, Corecursive g) => g -> g

regularEval :: forall f g. (Base g ~ f, BasicBase f, StuckBase f, AbortBase f, IndexedInputBase f, UnsizedBase f, Recursive g, Corecursive g, PrettyPrintable g) => g -> g

evalBU :: IExpr -> Either RunTimeError IExpr

evalBU' :: IExpr -> IO IExpr

term4toAbortExpr :: (Base g ~ f, BasicBase f, StuckBase f, AbortBase f, Corecursive g) => Term4 -> g

abortExprToTerm4 :: (Base g ~ f, BasicBase f, StuckBase f, AbortBase f, Foldable f, Recursive g) => g -> Either IExpr Term4

evalA :: StaticCheckExpr -> Maybe IExpr

evalPartial :: (Base g ~ f, Traversable f, BasicBase f, StuckBase f, DeferredEvalBase f, Recursive g, Corecursive g, PrettyPrintable g) => g -> g

evalPartial' :: IExpr -> IExpr

evalPartialUnsized :: Set Integer -> InputSizingExpr -> SizedRecursion

fullyResolve :: (Int -> Maybe PartialType) -> PartialType -> PartialType

buildTypeMap :: Set TypeAssociation -> Either TypeCheckError (Map Int PartialType)

partiallyAnnotate :: (Base g ~ f, Annotatable1 f, Annotatable g) => g -> Either TypeCheckError (PartialType, Int -> Maybe PartialType)

annotateTree :: forall g f. (Base g ~ f, Traversable f, Annotatable1 f, Recursive g, Annotatable g) => g -> Either TypeCheckError (Cofree f PartialType)

matchType :: PartialType -> PartialType -> Bool

matchTypeHead :: (Annotatable1 f, Annotatable g) => PartialType -> f g -> Bool

class GenValidAdj g where

Associated Types

type Gennable g

Methods

toGennable :: g -> Gennable g

fromGennable :: Gennable g -> g

genTypedTree :: Maybe PartialType -> PartialType -> Int -> Gen (Cofree UnsizedExprF PartialType)

tcAnnotatedProp :: Cofree UnsizedExprF PartialType -> Bool

Orphan instances

GenValid UnsizedRecursionToken

Instance details

Methods

genValid :: Gen UnsizedRecursionToken

shrinkValid :: UnsizedRecursionToken -> [UnsizedRecursionToken]

GenValid FunctionIndex

Instance details

Methods

genValid :: Gen FunctionIndex

shrinkValid :: FunctionIndex -> [FunctionIndex]

GenValid SizedRecursion

Instance details

Methods

genValid :: Gen SizedRecursion

shrinkValid :: SizedRecursion -> [SizedRecursion]

AbstractRunTime CompiledExpr

Instance details

Methods

eval :: CompiledExpr -> Either RunTimeError CompiledExpr

GenValid a => GenValid (AbortableF a)

Instance details

Methods

genValid :: Gen (AbortableF a)

shrinkValid :: AbortableF a -> [AbortableF a]

GenValid a => GenValid (IndexedInputF a)

Instance details

Methods

genValid :: Gen (IndexedInputF a)

shrinkValid :: IndexedInputF a -> [IndexedInputF a]

GenValid a => GenValid (PartExprF a)

Instance details

Methods

genValid :: Gen (PartExprF a)

shrinkValid :: PartExprF a -> [PartExprF a]

GenValid a => GenValid (StuckF a)

Instance details

Methods

genValid :: Gen (StuckF a)

shrinkValid :: StuckF a -> [StuckF a]

GenValid a => GenValid (SuperPositionF a)

Instance details

Methods

genValid :: Gen (SuperPositionF a)

shrinkValid :: SuperPositionF a -> [SuperPositionF a]

GenValid a => GenValid (UnsizedExprF a)

Instance details

Methods

genValid :: Gen (UnsizedExprF a)

shrinkValid :: UnsizedExprF a -> [UnsizedExprF a]

GenValid a => GenValid (UnsizedRecursionF a)

Instance details

Methods

genValid :: Gen (UnsizedRecursionF a)

shrinkValid :: UnsizedRecursionF a -> [UnsizedRecursionF a]

Validity a => Validity (AbortableF a)

Instance details

Methods

validate :: AbortableF a -> Validation

Validity a => Validity (IndexedInputF a)

Instance details

Methods

validate :: IndexedInputF a -> Validation

Validity a => Validity (PartExprF a)

Instance details

Methods

validate :: PartExprF a -> Validation

Validity a => Validity (StuckF a)

Instance details

Methods

validate :: StuckF a -> Validation

Validity a => Validity (SuperPositionF a)

Instance details

Methods

validate :: SuperPositionF a -> Validation

Validity a => Validity (UnsizedExprF a)

Instance details

Methods

validate :: UnsizedExprF a -> Validation

Validity a => Validity (UnsizedRecursionF a)

Instance details

Methods

validate :: UnsizedRecursionF a -> Validation

GenValid (Cofree UnsizedExprF PartialType)

Instance details

Methods

genValid :: Gen (Cofree UnsizedExprF PartialType)

shrinkValid :: Cofree UnsizedExprF PartialType -> [Cofree UnsizedExprF PartialType]

Validity (Cofree UnsizedExprF PartialType)

Instance details

Methods

validate :: Cofree UnsizedExprF PartialType -> Validation