Telomare.Resolver

Documentation

debug :: Bool

debugTrace :: String -> a -> a

type VarList = [String]

i2t :: LocTag -> Int -> Cofree (ParserTermF l v) LocTag

Int to ParserTerm

ints2t :: LocTag -> [Int] -> Cofree (ParserTermF l v) LocTag

List of Int's to ParserTerm

s2t :: LocTag -> String -> Cofree (ParserTermF l v) LocTag

String to ParserTerm

findInts :: LocTag -> Pattern -> [AnnotatedUPT -> AnnotatedUPT]

Finds all PatternInt leaves returning "directions" to these leaves through pairs in the form of a combination of RightUP and LeftUP from the root e.g. PatternPair (PatternVar "x") (PatternPair (PatternInt 0) (PatternVar "y")) will return [LeftUP . RightUP]

findStrings :: LocTag -> Pattern -> [AnnotatedUPT -> AnnotatedUPT]

Finds all PatternString leaves returning "directions" to these leaves through pairs in the form of a combination of RightUP and LeftUP from the root e.g. PatternPair (PatternVar "x") (PatternPair (PatternString "Hello, world!") (PatternVar "y")) will return [LeftUP . RightUP]

fitPatternVarsToCasedUPT :: Pattern -> AnnotatedUPT -> AnnotatedUPT

varsUPT :: UnprocessedParsedTerm -> Set String

Collect all free variable names in a AnnotatedUPT expresion

mkLambda4FreeVarUPs :: AnnotatedUPT -> AnnotatedUPT

findPatternVars :: LocTag -> Pattern -> Map String (AnnotatedUPT -> AnnotatedUPT)

pairStructureCheck :: Pattern -> UnprocessedParsedTerm -> UnprocessedParsedTerm

pairRoute2Dirs :: Pattern -> [UnprocessedParsedTerm -> UnprocessedParsedTerm]

pattern2UPT :: LocTag -> Pattern -> AnnotatedUPT

mkCaseAlternative

Arguments

:: AnnotatedUPT
UPT to be cased

-> AnnotatedUPT
Case result to be made lambda and applied

-> Pattern
Pattern

-> AnnotatedUPT
case result as a lambda applied to the appropirate part of the UPT to be cased

case2annidatedIfs

Arguments

:: AnnotatedUPT
Term to be pattern matched

-> [Pattern]
All patterns in a case expression

-> [AnnotatedUPT]
Int leaves as ListUPs on UPT

-> [AnnotatedUPT]
Int leaves as ListUPs on pattern

-> [AnnotatedUPT]
String leaves as ListUPs on UPT

-> [AnnotatedUPT]
String leaves as ListUPs on pattern

-> [AnnotatedUPT]
Case's alternatives

-> AnnotatedUPT

removeCaseUPs :: AnnotatedUPT -> AnnotatedUPT

debruijinize :: MonadFail m => VarList -> Term1 -> m Term2

rewriteOuterTag :: anno -> Cofree a anno -> Cofree a anno

splitExpr' :: Term2 -> BreakState' RecursionPieceFrag UnsizedRecursionToken

newtype FragExprUR'

Constructors

FragExprUR'

Fields

unFragExprUR' :: FragExpr (RecursionSimulationPieces FragExprUR')

newtype Term3'

Constructors

Term3' (Map FragIndex FragExprUR')

splitExpr :: Term2 -> Term3

makeLambda

Arguments

:: String
Variable name

-> Term1
Lambda body

-> Term1

`makeLambda ps vl t1` makes a TLam around t1 with vl as arguments. Automatic recognition of Close or Open type of TLam.

validateVariables :: AnnotatedUPT -> Either String Term1

Transformation from AnnotatedUPT to Term1 validating and inlining VarUPs

varsTerm1 :: Term1 -> Set String

Collect all free variable names in a Term1 expresion

optimizeBuiltinFunctions :: AnnotatedUPT -> AnnotatedUPT

generateAllHashes :: Term2 -> Term2

Process an Term2 to have all HashUP replaced by a unique number. The unique number is constructed by doing a SHA1 hash of the Term2 and adding one for all consecutive HashUP's.

addBuiltins :: AnnotatedUPT -> AnnotatedUPT

process :: AnnotatedUPT -> Either String Term3

Process an AnnotatedUPT to a Term3 with failing capability.

process2Term2 :: AnnotatedUPT -> Either String Term2

runTelomareParser2Term2

Arguments

:: TelomareParser AnnotatedUPT
Parser to run

-> String
Raw string to be parsed

-> Either String Term2
Error on Left

Helper function to compile to Term2

resolveImports'

Arguments

:: [(String, [Either AnnotatedUPT (String, AnnotatedUPT)])]

-> [Either AnnotatedUPT (String, AnnotatedUPT)]
Main module with both Import and Assignment

-> [Either AnnotatedUPT (String, AnnotatedUPT)]

resolveAllImports' :: [(String, [Either AnnotatedUPT (String, AnnotatedUPT)])] -> [Either AnnotatedUPT (String, AnnotatedUPT)] -> [Either AnnotatedUPT (String, AnnotatedUPT)]

resolveAllImports :: [(String, [Either AnnotatedUPT (String, AnnotatedUPT)])] -> [Either AnnotatedUPT (String, AnnotatedUPT)] -> [(String, AnnotatedUPT)]

resolveImports :: [(String, [Either AnnotatedUPT (String, AnnotatedUPT)])] -> String -> [(String, AnnotatedUPT)]

resolveMain

Arguments

:: [(String, [Either AnnotatedUPT (String, AnnotatedUPT)])]
Modules: [(ModuleName, [Either Import (VariableName, BindedUPT)])]

-> String
Module name with main

-> Either String AnnotatedUPT

main2Term3

Arguments

:: [(String, [Either AnnotatedUPT (String, AnnotatedUPT)])]
Modules: [(ModuleName, [Either Import (VariableName, BindedUPT)])]

-> String
Module name with main

-> Either String Term3
Error on Left

Orphan instances

MonadFail (Either String)
Int to Church encoding

Instance details

Methods

fail :: String -> Either String a