Telomare

Documentation

class BasicBase g where

Methods

embedB :: BasicExprF x -> g x

extractB :: g x -> Maybe (BasicExprF x)

class StuckBase g where

Methods

embedS :: StuckF x -> g x

extractS :: g x -> Maybe (StuckF x)

class AbortBase g where

Methods

embedA :: AbortableF x -> g x

extractA :: g x -> Maybe (AbortableF x)

pattern BasicFW :: BasicBase g => BasicExprF x -> g x

pattern BasicEE :: (Base g ~ f, BasicBase f, Recursive g) => BasicExprF g -> g

pattern StuckFW :: StuckBase g => StuckF x -> g x

pattern StuckEE :: (Base g ~ f, StuckBase f, Recursive g) => StuckF g -> g

pattern AbortFW :: AbortBase g => AbortableF x -> g x

pattern AbortEE :: (Base g ~ f, AbortBase f, Recursive g) => AbortableF g -> g

pattern ZeroB :: (Base g ~ f, BasicBase f, Recursive g) => g

pattern PairB :: (Base g ~ f, BasicBase f, Recursive g) => g -> g -> g

pattern FillFunction :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g) => g -> g -> f g

pattern GateSwitch :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g) => g -> g -> g -> f g

pattern AppEE :: (Base g ~ f, BasicBase f, StuckBase f, Recursive g) => g -> g -> g

basicEE :: (Base g ~ f, BasicBase f, Corecursive g) => BasicExprF g -> g

stuckEE :: (Base g ~ f, StuckBase f, Corecursive g) => StuckF g -> g

abortEE :: (Base g ~ f, AbortBase f, Corecursive g) => AbortableF g -> g

zeroB :: (Base g ~ f, BasicBase f, Corecursive g) => g

pairB :: (Base g ~ f, BasicBase f, Corecursive g) => g -> g -> g

envB :: (Base g ~ f, StuckBase f, Corecursive g) => g

setEnvB :: (Base g ~ f, StuckBase f, Corecursive g) => g -> g

gateB :: (Base g ~ f, StuckBase f, Corecursive g) => g -> g -> g

leftB :: (Base g ~ f, StuckBase f, Corecursive g) => g -> g

rightB :: (Base g ~ f, StuckBase f, Corecursive g) => g -> g

fillFunction :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g -> g -> g

gateSwitch :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g -> g -> g -> g

abortB :: (Base g ~ f, AbortBase f, Corecursive g) => g

firstArgB :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g

secondArgB :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g

thirdArgB :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g

fourthArgB :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g

fifthArgB :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g

varB :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => Int -> g

i2B :: (Base g ~ f, BasicBase f, Corecursive g) => Int -> g

b2i :: (Base g ~ f, BasicBase f, Recursive g) => g -> Maybe Int

b2s :: forall g f. (Base g ~ f, BasicBase f, Recursive g) => g -> Maybe String

s2b :: forall g f. (Base g ~ f, BasicBase f, Corecursive g) => String -> g

iteB_ :: (Base g ~ f, BasicBase f, StuckBase f, Corecursive g) => g -> g -> g -> g

data BasicExprF f

Constructors

ZeroSF

PairSF f f

type BasicExpr = Fix BasicExprF

data StuckF f

Constructors

DeferSF FunctionIndex f

EnvSF

SetEnvSF f

GateSF f f

LeftSF f

RightSF f

newtype FunctionIndex

Constructors

FunctionIndex

Fields

unFunctionIndex :: Int

data AbortableF f

Constructors

AbortF

AbortedF BasicExpr

data StuckExprF f

Constructors

StuckExprB (BasicExprF f)

StuckExprS (StuckF f)

type StuckExpr = Fix StuckExprF

data CompiledExprF f

Constructors

CompiledExprB (BasicExprF f)

CompiledExprS (StuckF f)

CompiledExprA (AbortableF f)

type CompiledExpr = Fix CompiledExprF

newtype UnsizedRecursionToken

Constructors

UnsizedRecursionToken

Fields

unUnsizedRecursionToken :: Int

data SourcePosition

Constructors

SourcePosition

Fields

sourcePositionLine :: Int

sourcePositionColumn :: Int

sourcePositionOffset :: Int

data SourceSpan

Constructors

SourceSpan

Fields

sourceSpanFile :: Maybe FilePath

sourceSpanStart :: SourcePosition

sourceSpanEnd :: SourcePosition

data LocTag

Constructors

SourceLoc SourceSpan

GeneratedLoc String (Maybe LocTag)

BuiltinLoc String

RuntimeLoc

DecompiledLoc

UnknownLoc

data Term3F f

Constructors

Term3B (BasicExprF f)

Term3S (StuckF f)

Term3A (AbortableF f)

Term3Unsized UnsizedRecursionToken

Term3CheckingWrapper LocTag f f

data LamType l

Lambdas can be closed if it's expresion does not depend on any outer binding.

Constructors

Open l

Closed l

LetBinding Int l

data ParserTerm l v

Parser AST

Constructors

TZero

TPair (ParserTerm l v) (ParserTerm l v)

TVar v

TApp (ParserTerm l v) (ParserTerm l v)

TCheck (ParserTerm l v) (ParserTerm l v)

TITE (ParserTerm l v) (ParserTerm l v) (ParserTerm l v)

TLeft (ParserTerm l v)

TRight (ParserTerm l v)

TTrace (ParserTerm l v)

THash (ParserTerm l v)

TChurch Int

TLam (LamType l) (ParserTerm l v)

TLimitedRecursion (ParserTerm l v) (ParserTerm l v) (ParserTerm l v)

TUnsizedRepeater

data ParserTermF (l :: Type) (v :: Type) r

Constructors

TZeroF

TPairF r r

TVarF v

TAppF r r

TCheckF r r

TITEF r r r

TLeftF r

TRightF r

TTraceF r

THashF r

TChurchF Int

TLamF (LamType l) r

TLimitedRecursionF r r r

TUnsizedRepeaterF

indent :: Int -> String -> String

Helper function to indent. Usefull for indented Show instances.

sindent :: String -> State Int String

Indentation with the State Monad.

indentWithOneChild :: String -> State Int String -> State Int String

One child indentation.

indentWithOneChild' :: String -> State Int String -> State Int String

indentWithTwoChildren' :: String -> State Int String -> State Int String -> State Int String

indentWithChildren' :: String -> [State Int String] -> State Int String

indentWithTwoChildren :: String -> State Int String -> State Int String -> State Int String

Two children indentation.

indentWithThreeChildren :: String -> State Int String -> State Int String -> State Int String -> State Int String

dropUntil :: (a -> Bool) -> [a] -> [a]

`dropUntil p xs` drops leading elements until `p $ head xs` is satisfied.

locStartLineColumn :: LocTag -> Maybe (Int, Int)

renderLocTag :: LocTag -> Maybe String

renderLocTagVerbose :: LocTag -> String

type Term1 = Cofree (ParserTermF String String) LocTag

type Term2 = Cofree (ParserTermF () Int) LocTag

type Term3 = Cofree Term3F LocTag

data TypeCheckError

Constructors

UnboundType Int

InconsistentTypes PartialType PartialType

RecursiveType Int

data ResolverError

Constructors

NoMainFunction String

ModuleNotFound String

DefinitionCycle [String]

MissingDefinitions [String]

MissingDefinitionAt LocTag String

ParseError String

renderResolverError :: ResolverError -> String

data EvalError

Constructors

RTE RunTimeError

TCE TypeCheckError

RE ResolverError

StaticCheckError String

CompileConversionError

RecursionLimitError UnsizedRecursionToken

data RunTimeError

Constructors

AbortRunTime BasicExpr

GenericRunTimeError String CompiledExpr

ResultConversionError String

class TelomareLike a where

Methods

fromTelomare :: StuckExpr -> a

toTelomare :: a -> Maybe StuckExpr

class TelomareLike a => AbstractRunTime a where

Methods

eval :: a -> Either RunTimeError a

type Term3Builder = State (FunctionIndex, UnsizedRecursionToken)

buildTerm :: Corecursive g => Term3Builder g -> g

deferS :: (Base g ~ f, StuckBase f, Corecursive g) => g -> Term3Builder g

pairS :: (Base g ~ CofreeF f a, BasicBase f, Recursive g, Corecursive g, Monad m) => m g -> m g -> m g

clamS :: forall g f. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g) => Term3Builder g -> Term3Builder g

lamS :: forall g f. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g) => Term3Builder g -> Term3Builder g

twiddleS :: forall g f. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Corecursive g) => Term3Builder g

appS :: forall g f. (Base g ~ CofreeF f LocTag, StuckBase f, BasicBase f, Recursive g, Corecursive g) => Term3Builder g -> Term3Builder g -> Term3Builder g

repeatFunctionS :: LocTag -> Term3Builder Term3

unsizedRepeater :: LocTag -> UnsizedRecursionToken -> Term3Builder Term3

repeaterAndAbort :: LocTag -> UnsizedRecursionToken -> Term3Builder Term3

i2CB :: LocTag -> Int -> Term3Builder Term3

unsizedRecursionWrapper :: LocTag -> Term3Builder Term3 -> Term3Builder Term3 -> Term3Builder Term3 -> Term3Builder Term3

data DataType

Constructors

ZeroType

ArrType DataType DataType

PairType DataType DataType

data PartialType

Constructors

ZeroTypeP

AnyType

TypeVariable LocTag Int

ArrTypeP PartialType PartialType

PairTypeP PartialType PartialType

toPartialType :: DataType -> PartialType

mergePairType :: DataType -> DataType

mergePairTypeP :: PartialType -> PartialType

containsFunction :: PartialType -> Bool

cleanType :: PartialType -> Bool

forget :: Corecursive a => Cofree (Base a) anno -> a

tag :: Recursive a => anno -> a -> Cofree (Base a) anno

convertBasic :: (BasicBase g, BasicBase h, Base x ~ h, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

convertStuck :: (StuckBase g, StuckBase h, Base x ~ h, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

convertAbort :: (AbortBase g, AbortBase h, Base x ~ h, Corecursive x, Monad m) => (g (m x) -> m x) -> g (m x) -> m x

insertAndGetKey :: (Ord e, Enum e) => a -> State (Map e a) e

pattern AbortRecursion :: (Base g ~ f, BasicBase f, Recursive g) => g -> g

pattern AbortUser :: (Base g ~ f, BasicBase f, Recursive g) => g -> g

pattern AbortAny :: (Base g ~ f, BasicBase f, Recursive g) => g

pattern AbortUnsizeable :: (Base g ~ f, BasicBase f, Recursive g) => g -> g

abortRecursion :: (Base g ~ f, BasicBase f, Corecursive g) => g -> g

abortUser :: (Base g ~ f, BasicBase f, Corecursive g) => g -> g

abortAny :: (Base g ~ f, BasicBase f, Corecursive g) => g

abortUnsizeable :: (Base g ~ f, BasicBase f, Corecursive g) => g -> g

convertAbortMessage :: (Base g ~ f, BasicBase f, Recursive g, Show g) => g -> String

data Pattern

AST for patterns in `case` expressions

Constructors

PatternVar String

PatternAnnotated Pattern UnprocessedParsedTerm

PatternInt Int

PatternString String

PatternIgnore

PatternPair Pattern Pattern

newtype LocatedName

Constructors

LocatedName (LocTag, String)

locatedNameLoc :: LocatedName -> LocTag

locatedNameText :: LocatedName -> String

locatedName :: LocTag -> String -> LocatedName

data UnprocessedParsedTerm

Firstly parsed AST sans location annotations

Constructors

VarUP String

ITEUP UnprocessedParsedTerm UnprocessedParsedTerm UnprocessedParsedTerm

LetUP [(LocatedName, UnprocessedParsedTerm)] UnprocessedParsedTerm

ListUP [UnprocessedParsedTerm]

IntUP Int

StringUP String

PairUP UnprocessedParsedTerm UnprocessedParsedTerm

AppUP UnprocessedParsedTerm UnprocessedParsedTerm

LamUP LocatedName UnprocessedParsedTerm

ChurchUP Int

UnsizedRecursionUP UnprocessedParsedTerm UnprocessedParsedTerm UnprocessedParsedTerm

LeftUP UnprocessedParsedTerm

RightUP UnprocessedParsedTerm

TraceUP UnprocessedParsedTerm

CheckUP UnprocessedParsedTerm UnprocessedParsedTerm

HashUP UnprocessedParsedTerm
On ad hoc user defined types, this term will be substitued to a unique Int.

UDTUP [String] UnprocessedParsedTerm

CaseUP UnprocessedParsedTerm [(Pattern, UnprocessedParsedTerm)]

ImportQualifiedUP String String

ImportUP String

data UnprocessedParsedTermF r

Constructors

VarUPF [Char]

ITEUPF r r r

LetUPF [(LocatedName, r)] r

ListUPF [r]

IntUPF Int

StringUPF [Char]

PairUPF r r

AppUPF r r

LamUPF LocatedName r

ChurchUPF Int

UnsizedRecursionUPF r r r

LeftUPF r

RightUPF r

TraceUPF r

CheckUPF r r

HashUPF r

UDTUPF [[Char]] r

CaseUPF r [(Pattern, r)]

ImportQualifiedUPF [Char] [Char]

ImportUPF [Char]

_ImportUP :: Prism' UnprocessedParsedTerm String

_ImportQualifiedUP :: Prism' UnprocessedParsedTerm (String, String)

_CaseUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, [(Pattern, UnprocessedParsedTerm)])

_UDTUP :: Prism' UnprocessedParsedTerm ([String], UnprocessedParsedTerm)

_HashUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_CheckUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_TraceUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_RightUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_LeftUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_UnsizedRecursionUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm, UnprocessedParsedTerm)

_ChurchUP :: Prism' UnprocessedParsedTerm Int

_LamUP :: Prism' UnprocessedParsedTerm (LocatedName, UnprocessedParsedTerm)

_AppUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_PairUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_StringUP :: Prism' UnprocessedParsedTerm String

_IntUP :: Prism' UnprocessedParsedTerm Int

_ListUP :: Prism' UnprocessedParsedTerm [UnprocessedParsedTerm]

_LetUP :: Prism' UnprocessedParsedTerm ([(LocatedName, UnprocessedParsedTerm)], UnprocessedParsedTerm)

_ITEUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm, UnprocessedParsedTerm)

_VarUP :: Prism' UnprocessedParsedTerm String

letBindingName :: (LocatedName, a) -> String

letBindingValue :: (LocatedName, a) -> a

letBindingLoc :: (LocatedName, a) -> LocTag

data PatternF r

Constructors

PatternVarF [Char]

PatternAnnotatedF r UnprocessedParsedTerm

PatternIntF Int

PatternStringF [Char]

PatternIgnoreF

PatternPairF r r